Incorporate CheckStyle into project
A new CheckStyle configuration file has been added to the project. The configuration file is located at `config/checkstyle/checkstyle.xml`. The configuration file is used to enforce a consistent coding style across the project. The configuration file is based on the Google Java Style Guide.
The CheckStyle configuration file is used to check the code for compliance with the Google Java Style Guide. The configuration file is used to enforce a consistent coding style across the project. The configuration file is based on the Google Java Style Guide.
Modifications have been made to adhere to the Google Java Style Guide.