Add GUI using JavaFX
A GUI has been added to the application using JavaFX.
The GUI is designed to be user-friendly and intuitive, allowing users to interact with the application more easily.
The GUI includes a main window that displays the user's tasks, a dialog box for user input, and a launcher class that initializes the application.
The GUI is designed to be responsive and interactive, providing users with a seamless experience when using the application.

Currently the GUI is half done, the main window is able to display the tasks and the dialog box is able to take in user input.
However, more decorations and features will be added in the future to enhance the user experience.